Force top-level client methods to be used internally
When the caching client calls it's super, the caching client method
calls should not be used internally. Currently it looks like with
how inheritance works in python the base class will still call the
sub class' methods when calling self. For now, muck with
self.__class__ to force the base methods where that's expected.